requirements.txt
playwright==1.40.0
openai==1.3.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
requests==2.31.0
lxml==4.9.3
